+++ pkg/cluster/internal/create/actions/createworker/azure.go
@@ -66,6 +66,26 @@ var azureCharts = ChartsDictionary{
 				"tigera-operator":      {Repository: "https://docs.projectcalico.org/charts", Version: "v3.30.2", Namespace: "tigera-operator", Pull: true, Reconcile: true},
 			},
 		},
+		"34": {
+			"managed": {},
+			"unmanaged": {
+				"azuredisk-csi-driver": {Repository: "https://raw.githubusercontent.com/kubernetes-sigs/azuredisk-csi-driver/master/charts", Namespace: "kube-system", Version: "1.33.5", Pull: false, Reconcile: false},
+				"azurefile-csi-driver": {Repository: "https://raw.githubusercontent.com/kubernetes-sigs/azurefile-csi-driver/master/charts", Namespace: "kube-system", Version: "1.34.1", Pull: false, Reconcile: false},
+				"cloud-provider-azure": {Repository: "https://raw.githubusercontent.com/kubernetes-sigs/cloud-provider-azure/master/helm/repo", Namespace: "kube-system", Version: "1.34.2", Pull: true, Reconcile: true},
+				"cluster-autoscaler":   {Repository: "https://kubernetes.github.io/autoscaler", Version: "9.52.1", Namespace: "kube-system", Pull: false, Reconcile: false},
+				"tigera-operator":      {Repository: "https://docs.projectcalico.org/charts", Version: "v3.31.5", Namespace: "tigera-operator", Pull: true, Reconcile: true},
+			},
+		},
+		"35": {
+			"managed": {},
+			"unmanaged": {
+				"azuredisk-csi-driver": {Repository: "https://raw.githubusercontent.com/kubernetes-sigs/azuredisk-csi-driver/master/charts", Namespace: "kube-system", Version: "1.34.4", Pull: false, Reconcile: false},
+				"azurefile-csi-driver": {Repository: "https://raw.githubusercontent.com/kubernetes-sigs/azurefile-csi-driver/master/charts", Namespace: "kube-system", Version: "1.35.3", Pull: false, Reconcile: false},
+				"cloud-provider-azure": {Repository: "https://raw.githubusercontent.com/kubernetes-sigs/cloud-provider-azure/master/helm/repo", Namespace: "kube-system", Version: "1.35.3", Pull: true, Reconcile: true},
+				"cluster-autoscaler":   {Repository: "https://kubernetes.github.io/autoscaler", Version: "9.57.0", Namespace: "kube-system", Pull: false, Reconcile: false},
+				"tigera-operator":      {Repository: "https://docs.projectcalico.org/charts", Version: "v3.31.5", Namespace: "tigera-operator", Pull: true, Reconcile: true},
+			},
+		},
 	},
 }