+++ pkg/cluster/internal/create/actions/createworker/provider.go
@@ -245,6 +245,16 @@ var commonsCharts = ChartsDictionary{
 				"flux2":        {Repository: "https://fluxcd-community.github.io/helm-charts", Version: "2.17.2", Namespace: "kube-system", Pull: true, Reconcile: true},
 			},
 		},
+		"35": {
+			"managed": {
+				"cert-manager": {Repository: "https://charts.jetstack.io", Version: "v1.20.2", Namespace: "cert-manager", Pull: true, Reconcile: true},
+				"flux2":        {Repository: "https://fluxcd-community.github.io/helm-charts", Version: "2.18.4", Namespace: "kube-system", Pull: true, Reconcile: true},
+			},
+			"unmanaged": {
+				"cert-manager": {Repository: "https://charts.jetstack.io", Version: "v1.20.2", Namespace: "cert-manager", Pull: true, Reconcile: true},
+				"flux2":        {Repository: "https://fluxcd-community.github.io/helm-charts", Version: "2.18.4", Namespace: "kube-system", Pull: true, Reconcile: true},
+			},
+		},
 	},
 }